Rename doctools module to coretools
