+++ novelwriter/core/__init__.py
@@ -19,7 +19,7 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <https://www.gnu.org/licenses/>.
 """
 
-from novelwriter.core.doctools import DocMerger, DocSplitter
+from novelwriter.core.coretools import DocMerger, DocSplitter
 from novelwriter.core.document import NWDoc
 from novelwriter.core.index import countWords
 from novelwriter.core.project import NWProject
