+++ novelwriter/core/coretools.py
